tracks/ietf-wimse/requirements.txt
beautifulsoup4>=4.12.0
PyYAML>=6.0
requests>=2.31.0
